refactor: split time cells
