+++ src/components/time/TimeCells.tsx
@@ -2,43 +2,9 @@ import React from 'react';
 import './TimeCells.css';
 import TimeTableCell from './TimeTableCell';
 import { Time } from '../../app/redux';
+import TimeCell from './TimeCell';
 
-export const TimeHeaderCells = () => (
-  <>
-    {[7, 8, 9, 10, 11, 12, 1, 2, 3, 4, 5].map((hour) => (
-      <TimeTableCell key={hour} align="right">
-        {hour}:00
-      </TimeTableCell>
-    ))}
-  </>
-);
-
-class TimeCell extends React.Component<any> {
-  constructor(props: any) {
-    super(props);
-
-    this.logTime = this.logTime.bind(this);
-    this.removeTime = this.removeTime.bind(this);
-  }
-
-  logTime() {
-    this.props.logTime(this.props.taskId, this.props.timeSegment);
-  }
-
-  removeTime() {
-    this.props.removeTime(this.props.taskId, this.props.timeSegment);
-  }
-
-  render() {
-    if (this.props.logged) {
-      return <div className="increment logged" onClick={this.removeTime} />;
-    } else {
-      return <div className="increment" onClick={this.logTime} />;
-    }
-  }
-}
-
-class TimeBodyCells extends React.Component<any> {
+class TimeCellRow extends React.Component<any> {
   render() {
     const cells: JSX.Element[] = [];
     for (let hour = 0; hour < 11; hour++) {
@@ -73,4 +39,4 @@ class TimeBodyCells extends React.Component<any> {
   }
 }
 
-export default TimeBodyCells;
+export default TimeCellRow;
