+++ src/components/time/TimeHeaderCell.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import TimeTableCell from './TimeTableCell';
+
+export const TimeHeaderCells = () => (
+  <>
+    {[7, 8, 9, 10, 11, 12, 1, 2, 3, 4, 5].map((hour) => (
+      <TimeTableCell key={hour} align="right">
+        {hour}:00
+      </TimeTableCell>
+    ))}
+  </>
+);
